split header and side menu
